Move Autocomplete component to folder
Code organization to move the Autocomplete component and its supporting component to a folder.
