+++ packages/studio-base/src/components/Autocomplete/Autocomplete.stories.tsx
@@ -15,8 +15,9 @@ import { Meta, StoryFn, StoryObj } from "@storybook/react";
 import { fireEvent, within } from "@storybook/testing-library";
 import * as _ from "lodash-es";
 
-import Autocomplete from "@foxglove/studio-base/components/Autocomplete";
 import Stack from "@foxglove/studio-base/components/Stack";
+
+import { Autocomplete } from "./Autocomplete";
 
 export default {
   title: "components/Autocomplete",
