+++ packages/studio-base/src/components/Autocomplete/ReactWindowListboxAdapter.tsx
